docs/requirements.txt
mkdocs>=1.5
mkdocs-material>=9.0
mkdocs-macros-plugin>=1.3
pandas>=2.2
pyyaml>=6.0
tabulate>=0.9